fixed according to the comments
